moved to api/utils
